rename example to example_sdtm_exporter
